Move definition pytest fixture methods to pytest.fixtures
